refactor(module/page): [auth] move components from components/smart to components & from page/forgot to page/auth/forgot & remove Auth, User Providers & use named export for forgot components
