+++ src/modules/auth/components/index.ts
@@ -1,1 +1,3 @@
-export * from './smart';
+export * from './Signin';
+export * from './Signup';
+export * from './Forgot';
