--- src/modules/auth/components/smart/index.ts
@@ -1,5 +1,0 @@
-export * from './Forgot';
-export * from './Signin';
-export * from './Signup';
-export * from './User';
-export * from './Auth';
