+++ src/pages/auth/forgot.tsx
@@ -1,0 +1,3 @@
+import { Forgot } from '@onr/auth';
+
+export default Forgot;
