--- src/pages/forgot.tsx
@@ -1,3 +1,0 @@
-import Forgot from '@onr/auth/components/smart/Forgot';
-
-export default Forgot;
